Add font property to various controls; Optimize most components to reflect new wiki; Bump to 1.2.0;

+++ ImGui.Forms/Controls/Label.cs
@@ -18,7 +18,7 @@ namespace ImGui.Forms.Controls
 
         public FontResource Font { get; set; }
 
-        public float LineDistance { get; set; }
+        public int LineDistance { get; set; }
 
         public ThemedColor TextColor { get; set; }
